requirements.txt
alembic==0.9.8
aniso8601==2.0.1
astroid==1.6.1
certifi==2018.1.18
click==6.7
cloudinary==1.11.0
Flask==0.12.2
Flask-Mail==0.9.0
Flask-HTTPAuth==3.2.3
Flask-Migrate==2.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-CORS==3.0.6
sightengine==1.4.0
gunicorn==19.7.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
mock==2.0.0
passlib==1.7.1
pbr==3.1.1
postgres==2.2.1
psycopg2==2.7.5
PyJWT==1.5.3
pylint==1.8.2
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2018.3
six==1.11.0
SQLAlchemy==1.2.5
urllib3==1.22
Werkzeug==0.14.1
wrapt==1.10.11
exponent-server-sdk==0.2.0
rollbar
google-api-python-client==1.7.9
marshmallow==2.19.5
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-language==1.2.0
google-cloud-vision==0.37.0
grpcio==1.22.0
flask-login==0.4.1
pytest==5.0.1